# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MARO - интерактивная карта с персонализированными маршрутами",
    "description": "Документация по использованию MARO. Для обращения к SignalR использовать: /api/maps",
    "version": "v1"
  },
  "paths": {
...
    "/api/Group/create_group": {
      "post": {
        "tags": [
          "Group"
        ],
        "summary": "Создать группу пользователей",
        "description": "SignalR результат: \"Notify\". Пример запроса:\r\n\r\n    POST: /api/group/create_group\r\n    {\r\n        \"userId\": \"4C2C522E-F785-4EB4-8ED7-260861453330\",\r\n        \"host\": \"http://example.com\"\r\n    }",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Удачно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGroupResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGroupResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGroupResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "У пользователя уже есть группа",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Роль не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Group/join_group": {
      "post": {
        "tags": [
          "Group"
        ],
        "summary": "Присоединиться к существующей группе",
        "description": "SignalR результат: \"Notify\". Пример запроса:\r\n\r\n    POST: /api/group/join_group\r\n    {\r\n        \"userId\": \"4C2C522E-F785-4EB4-8ED7-260861453330\",\r\n        \"groupId\": \"4C2C522E-F785-4EB4-8ED7-260861453330\"\r\n    }",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinTheGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinTheGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/JoinTheGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Удачно"
          },
          "400": {
            "description": "Поле UserId обязательно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Группа не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Group/delete_group/{groupId}": {
      "delete": {
        "tags": [
          "Group"
        ],
        "summary": "Удалить группу",
        "description": "Пример запроса:\r\n\r\n    DELETE: /api/group/delete_group/4C2C522E-F785-4EB4-8ED7-260861453330",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удачно"
          },
          "400": {
            "description": "Поле GroupId обязательно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Группа не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Group/group_details/{groupId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "summary": "Информация о группе",
        "description": "Пример запроса:\r\n\r\n    GET: /api/group/group_details/4C2C522E-F785-4EB4-8ED7-260861453330",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удачно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailsResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailsResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailsResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Поле GroupId обязательно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Группа не найдена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "host": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateGroupResponseModel": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "nullable": true
          },
          "qrLink": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupDetailsResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "qrLink": {
            "type": "string",
            "nullable": true
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "JoinTheGroupDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Insert JWT with Bearer into field",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}